added format of command line input
Still not sure what's the best way to prepare the inputs, as there could
be conflicts (i.e. when assigning reference different to the restart and
both are .rst files), so should predefine them in self.amber.bash. Still
need to run this code and clean up. Right now it should be able to find
'default' files, be assigned custom user files and write out all what is
needed to a directory.